Fix Windows build for cat and stat
