+++ commands/cat_appendmode_unix.go
@@ -1,0 +1,13 @@
+//go:build !windows
+
+package commands
+
+import "golang.org/x/sys/unix"
+
+func catHostAppendMode(file catHostHandle) bool {
+	flags, err := unix.FcntlInt(file.Fd(), unix.F_GETFL, 0)
+	if err != nil {
+		return false
+	}
+	return flags&unix.O_APPEND != 0
+}
