+++ commands/cat_appendmode_windows.go
@@ -1,0 +1,9 @@
+//go:build windows
+
+package commands
+
+func catHostAppendMode(file catHostHandle) bool {
+	// Windows host handles do not expose their open flags through the standard
+	// library, so fall back to non-append behavior for external redirections.
+	return false
+}
